Split AI video engines hub copy
